change size for loading image
StorageService.shared.downloadPresentImage()

+++ AuthViewFirebase/AuthViewFirebase/Screens/Home/Present/ViewModel/PresentViewModel.swift
@@ -215,7 +215,9 @@ class PresentModelViewModel: ObservableObject {
     // MARK: -- Удаляю подарок из коллекции "Wishlist"
     
     func removingPresentFromWishlist(_ idPresent: String) {
-        let docRef = Firestore.firestore().collection("Users").document(AuthService.shared.currentUser!.uid).collection("Wishlist").document(idPresent)
+        guard let user = currentUser else { return }
+        
+        let docRef = Firestore.firestore().collection("users").document(user.uid).collection("wishlist").document(idPresent)
         
         docRef.delete() { error in
             if let error = error {
